Added kmain ig

+++ kernel/include/math/conversion.h
@@ -1,9 +1,11 @@
 #pragma once
 
+#include <kernel.h>
+
 #define BYTE_TO_KB 1024.0
 #define KB_TO_MB 1024.0
 
-int bytesToKB(int bytes);
-int kbToMB(int kilobytes);
-int kbToBytes(int kilobytes);
-int mbToKB(int megabytes);
+double bytesToKB(double bytes);
+double kbToMB(double kilobytes);
+double kbToBytes(double kilobytes);
+double mbToKB(double megabytes);
